Finale

+++ database/migrations/2025_06_23_080201_create_appointments_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('appointments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('student_id');
+            $table->integer('student_id');
             $table->string('title');
             $table->date('appointment_date');
             $table->time('appointment_time');
